refactor: update tools to use pydantic models
